requirements.txt
aiogram==3.3.0
python-dotenv==1.0.1
asyncpg==0.29.0
redis==5.0.2
tortoise-orm==0.20.0
loguru==0.7.2